# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Dokumentasi Sederhana",
        "description": "Dokumentasi API untuk tabel statuses",
        "contact": {
            "email": "raden@example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/approvals": {
            "get": {
                "tags": [
                    "Approvals"
                ],
                "operationId": "82f0d0e97808a541b734a0b7614621e2",
                "responses": {
                    "200": {
                        "description": "List of approvals"
                    }
                }
            },
            "post": {
                "tags": [
                    "Approvals"
                ],
                "operationId": "a89973527e3af89764e0538a0b9b87fb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "expense_id",
                                    "approver_id",
                                    "status_id"
                                ],
                                "properties": {
                                    "expense_id": {
                                        "type": "integer"
                                    },
                                    "approver_id": {
                                        "type": "integer"
                                    },
                                    "status_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    }
                }
            }
        },
        "/api/v1/approvals/{id}": {
            "get": {
                "tags": [
                    "Approvals"
                ],
                "operationId": "8743c6bb2d2d0d306797d7be2282f857",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Approval detail"
                    }
                }
            },
            "put": {
                "tags": [
                    "Approvals"
                ],
                "operationId": "e56d960efbf6a79337114b41d24ff03b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "expense_id": {
                                        "type": "integer"
                                    },
                                    "approver_id": {
                                        "type": "integer"
                                    },
                                    "status_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Approvals"
                ],
                "operationId": "08dbc76a25379cbea0862debaca2a995",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted"
                    }
                }
            }
        },
        "/api/v1/approval-stages": {
            "get": {
                "tags": [
                    "ApprovalStages"
                ],
                "operationId": "957e0fb5633667d28c743de6042e88c9",
                "responses": {
                    "200": {
                        "description": "List of approval stages"
                    }
                }
            },
            "post": {
                "tags": [
                    "ApprovalStages"
                ],
                "operationId": "d37e806e3c187c1ad6216bcfc7a064ca",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "approver_id"
                                ],
                                "properties": {
                                    "approver_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    }
                }
            }
        },
        "/api/v1/approval-stages/{id}": {
            "get": {
                "tags": [
                    "ApprovalStages"
                ],
                "operationId": "823ce6bb468a04c5ae529e46eb67f4bf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Approval stage detail"
                    }
                }
            },
            "put": {
                "tags": [
                    "ApprovalStages"
                ],
                "operationId": "ce5367eedf661ff5ba7d5573175c0ceb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "approver_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated"
                    }
                }
            },
            "delete": {
                "tags": [
                    "ApprovalStages"
                ],
                "operationId": "0f076958e31ed100cbb728f2c24b4ae9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted"
                    }
                }
            }
        },
...
    },
...
}